Refactor: 상위 커스텀 예외를 common.exception으로 이동
