+++ src/main/java/com/uranus/taskmanager/api/authentication/exception/InvalidLoginIdentityException.java
@@ -2,6 +2,8 @@ package com.uranus.taskmanager.api.authentication.exception;
 
 import org.springframework.http.HttpStatus;
 
+import com.uranus.taskmanager.api.common.exception.AuthenticationException;
+
 public class InvalidLoginIdentityException extends AuthenticationException {
 
 	private static final String MESSAGE = "Please provide a valid login ID or Email.";
