+++ src/main/java/com/uranus/taskmanager/api/common/exception/AuthenticationException.java
@@ -1,8 +1,6 @@
-package com.uranus.taskmanager.api.authentication.exception;
+package com.uranus.taskmanager.api.common.exception;
 
 import org.springframework.http.HttpStatus;
-
-import com.uranus.taskmanager.api.common.exception.CommonException;
 
 public abstract class AuthenticationException extends CommonException {
