+++ src/main/java/com/uranus/taskmanager/api/common/exception/InvitationException.java
@@ -1,8 +1,6 @@
-package com.uranus.taskmanager.api.invitation.exception;
+package com.uranus.taskmanager.api.common.exception;
 
 import org.springframework.http.HttpStatus;
-
-import com.uranus.taskmanager.api.common.exception.CommonException;
 
 public class InvitationException extends CommonException {
