+++ src/main/java/com/uranus/taskmanager/api/common/exception/WorkspaceException.java
@@ -1,8 +1,6 @@
-package com.uranus.taskmanager.api.workspace.exception;
+package com.uranus.taskmanager.api.common.exception;
 
 import org.springframework.http.HttpStatus;
-
-import com.uranus.taskmanager.api.common.exception.CommonException;
 
 public abstract class WorkspaceException extends CommonException {
 	public WorkspaceException(String message, HttpStatus httpStatus) {
