+++ src/main/java/com/uranus/taskmanager/api/common/exception/WorkspaceMemberException.java
@@ -1,8 +1,6 @@
-package com.uranus.taskmanager.api.workspacemember.exception;
+package com.uranus.taskmanager.api.common.exception;
 
 import org.springframework.http.HttpStatus;
-
-import com.uranus.taskmanager.api.common.exception.CommonException;
 
 public abstract class WorkspaceMemberException extends CommonException {
