+++ src/main/java/com/uranus/taskmanager/api/invitation/exception/InvalidInvitationStatusException.java
@@ -2,6 +2,8 @@ package com.uranus.taskmanager.api.invitation.exception;
 
 import org.springframework.http.HttpStatus;
 
+import com.uranus.taskmanager.api.common.exception.InvitationException;
+
 public class InvalidInvitationStatusException extends InvitationException {
 	private static final String MESSAGE = "Invitation has already been processed";
 	private static final HttpStatus HTTP_STATUS = HttpStatus.BAD_REQUEST;
