+++ src/main/java/com/uranus/taskmanager/api/invitation/exception/InvitationAlreadyExistsException.java
@@ -2,6 +2,8 @@ package com.uranus.taskmanager.api.invitation.exception;
 
 import org.springframework.http.HttpStatus;
 
+import com.uranus.taskmanager.api.common.exception.InvitationException;
+
 public class InvitationAlreadyExistsException extends InvitationException {
 	private static final String MESSAGE = "An invitation for this member is already pending";
 	private static final HttpStatus HTTP_STATUS = HttpStatus.BAD_REQUEST;
