+++ src/main/java/com/uranus/taskmanager/api/member/exception/DuplicateLoginIdException.java
@@ -2,6 +2,8 @@ package com.uranus.taskmanager.api.member.exception;
 
 import org.springframework.http.HttpStatus;
 
+import com.uranus.taskmanager.api.common.exception.MemberException;
+
 public class DuplicateLoginIdException extends MemberException {
 
 	private static final String MESSAGE = "The given Login ID already exists";
