+++ src/main/java/com/uranus/taskmanager/api/member/exception/MemberNotFoundException.java
@@ -2,6 +2,8 @@ package com.uranus.taskmanager.api.member.exception;
 
 import org.springframework.http.HttpStatus;
 
+import com.uranus.taskmanager.api.common.exception.MemberException;
+
 public class MemberNotFoundException extends MemberException {
 
 	private static final String MESSAGE = "Member was not found";
