+++ src/main/java/com/uranus/taskmanager/api/workspace/exception/WorkspaceNotFoundException.java
@@ -2,6 +2,8 @@ package com.uranus.taskmanager.api.workspace.exception;
 
 import org.springframework.http.HttpStatus;
 
+import com.uranus.taskmanager.api.common.exception.WorkspaceException;
+
 public class WorkspaceNotFoundException extends WorkspaceException {
 
 	private static final String MESSAGE = "Workspace was not found for the given code";
