+++ src/main/java/com/uranus/taskmanager/api/workspacemember/exception/MemberAlreadyParticipatingException.java
@@ -2,6 +2,8 @@ package com.uranus.taskmanager.api.workspacemember.exception;
 
 import org.springframework.http.HttpStatus;
 
+import com.uranus.taskmanager.api.common.exception.WorkspaceMemberException;
+
 public class MemberAlreadyParticipatingException extends WorkspaceMemberException {
 	private static final String MESSAGE = "Current login member is already participating in this Workspace";
 	private static final HttpStatus HTTP_STATUS = HttpStatus.BAD_REQUEST;
