+++ src/main/java/com/uranus/taskmanager/api/workspacemember/exception/MemberNotInWorkspaceException.java
@@ -2,6 +2,8 @@ package com.uranus.taskmanager.api.workspacemember.exception;
 
 import org.springframework.http.HttpStatus;
 
+import com.uranus.taskmanager.api.common.exception.WorkspaceMemberException;
+
 public class MemberNotInWorkspaceException extends WorkspaceMemberException {
 	private static final String MESSAGE = "Member was not found in the given workspace";
 	private static final HttpStatus HTTP_STATUS = HttpStatus.NOT_FOUND;
